Fixed directory deletion
directories with files in them were not being deleted correctly. Needed to use new package `del` to allow for easier file deletion.
