+++ package.json
@@ -40,6 +40,7 @@
         "@types/node": "^12.6.2",
         "@types/uuid": "^3.4.5",
         "chai": "^4.2.0",
+        "del": "^5.0.0",
         "mocha": "^6.1.4",
         "nodemon": "^1.19.1",
         "ts-node": "^8.3.0",
