+++ tsconfig.json
@@ -1,7 +1,6 @@
 {
 	"compilerOptions": {
 		"module": "commonjs",
-		"baseUrl": "./src",
 		"outDir": "./out/",
 		"sourceMap": false,
 		"typeRoots": [
